requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-groq==0.0.1
pydantic==2.6.1
requests>=2.31.0
langsmith>=0.0.63,<0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.13
langgraph>=0.0.15
langchain-huggingface==0.0.2
google-generativeai==0.3.2
yfinance==0.2.31
numpy==1.26.2
pandas==2.2.0
scikit-learn==1.6.1
joblib==1.3.2
pinecone-client==2.2.4
huggingface-hub>=0.23.0,<0.24.0
tavily-python==0.2.6
typing-extensions==4.8.0
rich>=13.7.0
flask==3.0.2
python-multipart==0.0.6
httpx==0.25.1 